Create FlyController.php
I so far created a FlyController and make six functions in it with index being listing what’s in the migrations. Creates function will be an add form then a store function that will submit the model to the database. Also a edit function form that submits to a update function which will update any changes. Lastly a destroy function that will take delete.